migrate tests to single code base
