+++ tests/integration/conftest.py
@@ -2,10 +2,12 @@
 # See LICENSE file for licensing details.
 
 import logging
+from platform import machine
 
 import jubilant
 import pytest
 
+from literals import Substrate
 from tests.integration.continuous_writes import ContinuousWrites
 from tests.integration.helpers import APP_NAME
 
@@ -27,3 +29,37 @@ def c_writes_runner(juju: jubilant.Juju, c_writes: ContinuousWrites):
     yield
     logger.info("Clearing continuous writes after test completion")
     logger.info(c_writes.clear())
+
+
+@pytest.fixture(scope="session")
+def substrate(request) -> Substrate:
+    """Substrate that we are testing."""
+    return Substrate(request.config.option.substrate)
+
+
+@pytest.fixture(scope="package")
+def arch() -> str:
+    """Fixture to provide the platform architecture for testing."""
+    platforms = {
+        "x86_64": "amd64",
+        "aarch64": "arm64",
+    }
+    return platforms.get(machine(), "amd64")
+
+
+@pytest.fixture
+def charm(arch: str) -> str:
+    """Path to the charm file to use for testing."""
+    # Return str instead of pathlib.Path since python-libjuju's model.deploy(), juju deploy, and
+    # juju bundle files expect local charms to begin with `./` or `/` to distinguish them from
+    # Charmhub charms.
+    return f"./valkey_ubuntu@24.04-{arch}.charm"
+
+
+@pytest.fixture(scope="module")
+def juju(arch: str):
+    # `testing` is the default model created by concierge
+    juju = jubilant.Juju(model="testing")
+    juju.wait_timeout = 1000
+    juju.cli("set-model-constraints", f"arch={arch}")
+    yield juju
