+++ tests/integration/test_charm.py
@@ -9,6 +9,7 @@ import pytest
 from literals import (
     INTERNAL_USERS_PASSWORD_CONFIG,
     CharmUsers,
+    Substrate,
 )
 from statuses import CharmStatuses, ClusterStatuses
 from tests.integration.helpers import (
@@ -38,9 +39,14 @@ TEST_KEY = "test_key"
 TEST_VALUE = "test_value"
 
 
-def test_build_and_deploy(charm: str, juju: jubilant.Juju) -> None:
+def test_build_and_deploy(charm: str, juju: jubilant.Juju, substrate: Substrate) -> None:
     """Build the charm-under-test and deploy it with three units."""
-    juju.deploy(charm, resources=IMAGE_RESOURCE, num_units=NUM_UNITS, trust=True)
+    juju.deploy(
+        charm,
+        resources=IMAGE_RESOURCE if substrate == Substrate.K8S else None,
+        num_units=NUM_UNITS,
+        trust=True,
+    )
     juju.wait(
         lambda status: are_apps_active_and_agents_idle(status, APP_NAME, idle_period=30),
         timeout=600,
