+++ tests/spread/vm/test_charm.py/task.yaml
@@ -2,6 +2,6 @@ summary: test_charm.py
 environment:
   TEST_MODULE: test_charm.py
 execute: |
-  tox run -e integration -- "tests/integration/vm/$TEST_MODULE" --alluredir="$SPREAD_TASK/allure-results"
+  tox run -e integration -- "tests/integration/vm/$TEST_MODULE" --substrate vm --alluredir="$SPREAD_TASK/allure-results"
 artifacts:
   - allure-results
