+++ tests/spread/vm/test_scaling.py/task.yaml
@@ -2,6 +2,6 @@ summary: test_scaling.py
 environment:
   TEST_MODULE: ha/test_scaling.py
 execute: |
-  tox run -e integration -- "tests/integration/vm/$TEST_MODULE" --alluredir="$SPREAD_TASK/allure-results"
+  tox run -e integration -- "tests/integration/vm/$TEST_MODULE" --substrate vm --alluredir="$SPREAD_TASK/allure-results"
 artifacts:
   - allure-results
